+++ src/ULinkRPC.CodeGen/FacadeEmitter.cs
@@ -72,7 +72,7 @@ internal static partial class FacadeEmitter
 
         w.OpenBlock("public RpcClient(RpcClientOptions options)");
         w.Line("Options = options ?? throw new ArgumentNullException(nameof(options));");
-        w.Line("_runtime = new RpcClientRuntime(options.Transport, options.Serializer, options.KeepAlive);");
+        w.Line("_runtime = new RpcClientRuntime(options);");
         w.CloseBlock();
         w.Line();